+++ Bootstrap/Utils/Console.cpp
@@ -108,6 +108,8 @@ Console::Color Console::GetRainbowColor()
 
 void Console::SetColor(Color color)
 {
+	if (!IsInitialized())
+		return;
 	color = ((Mode == DisplayMode::MAGENTA)
 		? Color::Magenta
 		: (((Mode == DisplayMode::RAINBOW) || (Mode == DisplayMode::RANDOMRAINBOW))
@@ -116,53 +118,53 @@ void Console::SetColor(Color color)
 	switch (color)
 	{
 	case Color::Black:
-		Write("\x1b[30m");
+		std::cout << "\x1b[30m";
 		break;
 	case Color::DarkBlue:
-		Write("\x1b[34m");
+		std::cout << "\x1b[34m";
 		break;
 	case Color::DarkGreen:
-		Write("\x1b[32m");
+		std::cout << "\x1b[32m";
 		break;
 	case Color::DarkCyan:
-		Write("\x1b[36m");
+		std::cout << "\x1b[36m";
 		break;
 	case Color::DarkRed:
-		Write("\x1b[31m");
+		std::cout << "\x1b[31m";
 		break;
 	case Color::DarkMagenta:
-		Write("\x1b[35m");
+		std::cout << "\x1b[35m";
 		break;
 	case Color::DarkYellow:
-		Write("\x1b[33m");
+		std::cout << "\x1b[33m";
 		break;
 	case Color::Gray:
-		Write("\x1b[37m");
+		std::cout << "\x1b[37m";
 		break;
 	case Color::DarkGray:
-		Write("\x1b[90m");
+		std::cout << "\x1b[90m";
 		break;
 	case Color::Blue:
-		Write("\x1b[94m");
+		std::cout << "\x1b[94m";
 		break;
 	case Color::Green:
-		Write("\x1b[92m");
+		std::cout << "\x1b[92m";
 		break;
 	case Color::Cyan:
-		Write("\x1b[96m");
+		std::cout << "\x1b[96m";
 		break;
 	case Color::Red:
-		Write("\x1b[91m");
+		std::cout << "\x1b[91m";
 		break;
 	case Color::Magenta:
-		Write("\x1b[95m");
+		std::cout << "\x1b[95m";
 		break;
 	case Color::Yellow:
-		Write("\x1b[93m");
+		std::cout << "\x1b[93m";
 		break;
 	case Color::White:
 	default:
-		Write("\x1b[97m");
+		std::cout << "\x1b[97m";
 		break;
 	}
 }
@@ -171,5 +173,10 @@ void Console::Write(const char* txt)
 {
 	if (!IsInitialized())
 		return;
+	CONSOLE_SCREEN_BUFFER_INFO info;
+	if (!GetConsoleScreenBufferInfo(OutputHandle, &info))
+		return;
+	Color original = (Color)info.wAttributes;
 	std::cout << txt;
+	SetColor(original);
 };